+++ CHANGELOG.md
@@ -36,6 +36,7 @@
   - Add `--id` and `--search` options to `open` command to open given zettel ID or search page respectively  (#317)
   - Use current directory as the default value for `-d` (a la. git) (#389)
   - Adapt `neuron new` for title IDs (#408)
+  - `neuron search`: fail if org is enabled, unless --full-text is used. (#420)
 - Static binaries
 
 ## 0.6.0.0